small changes in orders pages
