+++ Frontend/src/App.tsx
@@ -11,7 +11,7 @@ import { UserProvider } from './context/UserContext.tsx';
 import { DateProvider } from './context/DateContext.tsx';
 import CourseTemplateDisplay from './components/Admin/Courses/CourseTemplates/CourseTemplateDisplay.tsx';
 import Dashboard from './components/Admin/Dashboard/Dashboard.tsx';
-import UserProfile from './components/Members/UserProfile.tsx';
+import UserProfile from './components/Members/Profile/UserProfile.tsx';
 import Orders from './components/Members/Orders/Orders.tsx';
 import AdminProducts from './components/Admin/Products/AdminProducts.tsx';
 import AdminProfile from './components/Admin/Profile/AdminProfile.tsx';
